01. Stock Price
THE PROBLEM:

Suppose we could access yesterday’s stock prices as an array, where:

* the indices are the time in minutes past trade opening time, which
  was 9:30 AM local time
* the values are the price in dollars of stock at that time

So if the stock cost $500 at 10:30 AM, stockPricesYesterday[60] = 500

Write an efficient function that takes stockPricesYesterday and
returns the best profit you can make from 1 purchase and 1 sale of
Apple stock yesterday.

No “shorting”. You must buy before you sell. You may not guy and sell in
the same time step (at least 1 minute must pass).

SOLUTION:

In this initial solution, I attempted to base the solution off of the
max stock price in the array (when the index is greater than zero).
The algorithm I designed looked like this:

What is the maximum value in the array?
If the maximum value is greater than zero,
  take the set of all prices with an index less than the max
  find the minimum value in the set
  calculate profit by subtracting the minimum previous price from the
  max price
Otherwise
  the max price is the first price in the array
  take the set of all remaining prices
  find the maximum value in the set
  calculate min loss by subtracting the maximum sell price from the
  first price in the array

There may be a case when the best profit calculation is not based on
the maximum price in the array. This solution does not take those
cases into account.