01. Stock Price - Optimal Solution
This solution is in O(n) time and O(1) space.

Where I went wrong in my thinking about this problem is that I wanted
to be able to arrive at an answer based on the maximum price in the
array. This may in some cases arrive at the appropriate solution, but
there can be cases where the difference between the lowest price after
the maximum value and a higher value result in a higher profit.

So, you do have to iterate through the array at least once and track
the maximum profit by comparing the profit you would derive using the
current value in the index to all prior potential profits.

This has O(n) time complexity because the time it takes in order to
arrive at a solution grows with the size of the array. We only have to
traverse the array one time in order to arrive at the solution.